refactor: Coordinator와 ViewModel간의 의존성 분리
