+++ LeaveGo/LeaveGo/Shared/Models/PlaceDTO.swift
@@ -77,4 +77,3 @@ extension PlaceDTO {
         )
     }
 }
-
